[Bug] Fix out-of-domain coloring for valid strings
deck.gl BinSorter only properly sorts numbers. It uses '-' as it's
comparison operator. This operator returns NaN for strings which
causes the domain to be sorted randomly. We then decide if a
value is valid by comparing to the first and last elements of the
"sorted" domain which do not provide accurate bounds for string
elements. Short of making BinSorter work for multiple types, the
easiest fix is to check if a string value is in the valueDomain
of the dimensonState.